Split admin logic between three, not two pages.
Logically separated admin functionality into three admin pages,
register, virtallotteryregister & submitlottery. The separating is
actions to be done before, during & after the lottery.

Also did some styling cleanup.
